Split indicator utilities into database, datatable
