+++ src/pages/dataTableUtilities/getIndicatorTableData.js
@@ -4,7 +4,7 @@
  */
 
 import { COLUMN_TYPES } from './constants';
-import { UIDatabase, createRecord } from '../../database';
+import { getRowColumnIndicatorValue } from '../../database/utilities/getIndicatorData';
 
 const COLUMN_INDICATOR = {
   width: 1,
@@ -37,119 +37,6 @@ const COLUMNS = {
 };
 
 /**
- * Returns a filter method for finding a subset of an array of IndicatorAttribute objects by period.
- * @param {Array.<IndicatorAttribute>} filterPeriod
- * @return {Func}
- */
-const filterByPeriod = filterPeriod => ({ period }) => period.id === filterPeriod.id;
-
-/**
- * Returns a find method for querying an array of indicatorValue for the index of a given object.
- * @param {IndicatorValue} findValue
- * @return {Func}
- */
-const findByValue = findValue => value => findValue.id === value.id;
-
-/**
- * Query if an IndicatorValue array contains a given object.
- * @param {Array.<IndicatorValue>} values
- * @param {IndicatorValue} value
- */
-const includesValue = (values, value) => {
-  const result = values.findIndex(findByValue(value)) > 0;
-  return result;
-};
-
-/**
- * Returns a filter method for getting the intersection of two arrays of IndicatorValue objects.
- * @param {Array.<IndicatorValue>} values
- * @return {Func}
- */
-const filterByValues = values => value => includesValue(values, value);
-
-/**
- * Create a new indicator value.
- * @param {IndicatorAttribute} row
- * @param {IndicatorAttribute} column
- * @param {Period} period
- * @return {IndicatorValue}
- */
-const createIndicatorValue = (row, column, period) => {
-  let record;
-  UIDatabase.write(() => {
-    record = createRecord(UIDatabase, 'IndicatorValue', row, column, period);
-  });
-  return record;
-};
-
-/**
- * Update indicator value object with new value.
- * @param {IndicatorValue} indicatorValue
- * @param {string} value
- */
-const updateIndicatorValue = (indicatorValue, value) => {
-  UIDatabase.write(() => {
-    UIDatabase.update('IndicatorValue', { ...indicatorValue, value });
-  });
-};
-
-/**
- * Get value for indicator row, column attribute pair.
- *
- * @param {IndicatorAttribute} row
- * @param {IndicatorAttribute} column
- * @param {Period} period
- * @return {IndicatorValue}
- */
-const getIndicatorValue = (row, column, period) => {
-  const rowValues = row.values.filter(filterByPeriod(period));
-  const columnValues = column.values.filter(filterByPeriod(period));
-  const rowColumnValues = rowValues?.filter(filterByValues(columnValues));
-  const [rowColumnValue] = rowColumnValues;
-  return rowColumnValue || createIndicatorValue(row, column, period);
-};
-
-/**
- * Find indicator row by id.
- * @param {Array.<IndicatorAttribute} indicatorRows
- * @param {string} rowId
- * @return {IndicatorAttribute}
- */
-const getIndicatorRow = (indicatorRows, rowId) => indicatorRows.find(({ id }) => id === rowId);
-
-/**
- * Get indicator rows by id.
- * @param {ProgramIndicator} indicator
- * @param {string} id
- * @return {Array.<IndicatorAttribute>}
- */
-const getIndicatorRows = (indicator, id) =>
-  indicator?.rows?.filter(({ id: rowId }) => rowId === id);
-
-/**
- * Find indicator column by code.
- * @param {Array.<IndicatorAttribute>} indicatorColumns
- * @param {string} columnCode
- * @return {IndicatorAttribute}
- */
-const getIndicatorColumn = (indicatorColumns, columnCode) =>
-  indicatorColumns.find(({ code }) => code === columnCode);
-
-/**
- * Get indicator columns by code.
- * @param {ProgramIndicator} indicator
- * @param {string} code
- * @return {Array.<IndicatorAttribute>}
- */
-const getIndicatorColumns = (indicator, code) =>
-  indicator?.columns?.filter(({ code: columnCode }) => columnCode === code);
-
-const getIndicatorData = indicator => ({
-  rows: indicator.rows,
-  columns: indicator.columns,
-});
-
-/**
  * Get data table row for indicator row attribute.
  * @param {IndicatorAttribute} row
  * @param {ProgramIndicator} indicator
@@ -160,7 +47,7 @@ const mapIndicatorTableRow = (row, period) => {
   const { id, description, code, indicator } = row;
   const values = indicator.columns.reduce((acc, column) => {
     const { code: key } = column;
-    const value = getIndicatorValue(row, column, period);
+    const value = getRowColumnIndicatorValue(row, column, period);
     return { ...acc, [key]: value.value };
   }, {});
   return { id, description, code, ...values };
@@ -172,7 +59,12 @@ const mapIndicatorTableRow = (row, period) => {
  * @param {Period} period
  * @return {Array.<object>}
  */
-const mapIndicatorTableRows = (rows, period) => rows.map(row => mapIndicatorTableRow(row, period));
+const mapIndicatorTableRows = (rows, period) => {
+  console.log(period);
+  const tableRows = rows.map(row => mapIndicatorTableRow(row, period));
+  console.log(tableRows);
+  return tableRows;
+};
 
 /**
  * Get indicator data table row objects.
@@ -190,10 +82,11 @@ const getIndicatorTableRows = (indicator, period) => {
  * @param {IndicatorAttribute} column
  * @return {object}
  */
-const mapIndicatorTableColumn = column => {
+const mapIndicatorTableColumn = (column, isEditable) => {
+  const tableColumn = isEditable ? COLUMN_INDICATOR_MUTABLE : COLUMN_INDICATOR_IMMUTABLE;
   const { description: title, code: key } = column;
   return {
-    ...COLUMN_INDICATOR_MUTABLE,
+    ...tableColumn,
     title,
     key,
   };
@@ -202,10 +95,11 @@ const mapIndicatorTableColumn = column => {
 /**
  * Map indicator columns to data table column objects.
  * @param {Array.<IndicatorAttribute>} indicatorColumns
+ * @oaram {bool} isEditable
  * @return {Array.<object>}
  */
-const mapIndicatorTableColumns = indicatorColumns => {
-  const valueColumns = indicatorColumns.map(mapIndicatorTableColumn);
+const mapIndicatorTableColumns = (indicatorColumns, isEditable) => {
+  const valueColumns = indicatorColumns.map(column => mapIndicatorTableColumn(column, isEditable));
   return [COLUMNS.DESCRIPTION, COLUMNS.CODE, ...valueColumns];
 };
 
@@ -232,14 +126,6 @@ const getIndicatorTableData = (indicator, period) => {
 };
 
 export {
-  createIndicatorValue,
-  updateIndicatorValue,
-  getIndicatorValue,
-  getIndicatorRow,
-  getIndicatorRows,
-  getIndicatorColumn,
-  getIndicatorColumns,
-  getIndicatorData,
   mapIndicatorTableRow,
   mapIndicatorTableRows,
   getIndicatorTableRows,
